split function, manager and model manager and moved to weak ref datastructures
