+++ hyperFakeWorker/worker/function/model.py
@@ -1,68 +1,21 @@
-from dataclasses import dataclass
+from pathlib import Path
 
-import numpy as np
-from ..utils.onnx import ONNXModelInferer
+from .image import FunctionImage
 
-@dataclass
-class FunctionModelInput():
-    body_size: int
-    function_instances: int
-    function_calls: int
-    cpu_usage: float
-    ram_usage: int
+from . import FunctionIdStr
 
-    def as_tensor(self) -> np.ndarray:
-        return np.array(
-            [[
-                self.body_size,
-                self.function_instances,
-                self.function_calls,
-                self.cpu_usage,
-                self.ram_usage
-            ]],
-            dtype=np.float32
-        )
+class ModelManager():
 
-@dataclass
-class FunctionModelOutput():
-    function_runtime: int
-    cpu_usage: float
-    ram_usage: int
+    def __init__(self, models: list[Path]):
 
-    @staticmethod
-    def from_tensor(tensor: np.ndarray) -> "FunctionModelOutput":
-        values = tensor.flatten()
+        self.function_model_paths = models
+        self.function_models: dict[FunctionIdStr, Path] = {}
 
-        if values.size != 3:
-            raise ValueError(f"Expected 3 otuput values, got {values.size}: {values}")
-
-        return FunctionModelOutput(
-            function_runtime=int(values[0]),
-            cpu_usage=float(values[1]),
-            ram_usage=int(values[2])
-        )
-
-class FunctionModelInferer(ONNXModelInferer):
-
-    def infer(self, input_data: FunctionModelInput) -> FunctionModelOutput:
-        """
-        Perform inference using the loaded ONNX model.
-        Args:
-            input_data (FunctionModelInput): A FunctionModelInput object
-        Returns:
-            FunctionModelOutput: The output of the model inference.
-        """
-        tensor_input = input_data.as_tensor()
-
-        input_names = self.model.get_inputs()
-
-        if len(input_names) != 1:
-            raise ValueError(f"Expected exactly one input in the model, got {len(input_names)}.")
-
-        input_name = input_names[0].name
-
-        outputs = self.model.run(None, {input_name: tensor_input})
-
-        output_tensor = outputs[0]
-
-        return FunctionModelOutput.from_tensor(output_tensor)
+    def find_model(self, function_id: FunctionIdStr, image: FunctionImage) -> Path:
+        # resolve model path
+        if self.function_models.get(function_id) is None:
+            for p in self.function_model_paths:
+                if p.stem == image.image:
+                    self.function_models[function_id] = p
+                    break
+        return self.function_models.get(function_id)
